Use int and uint32_t instead of long and unsigned long, when appropriate.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@23452 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ mp3lib/decode_MMX.c
@@ -14,7 +14,7 @@
 extern void (*dct64_MMX_func)(short*, short*, real*);
 static unsigned long long attribute_used __attribute__((aligned(8))) null_one = 0x0000ffff0000ffffULL;
 static unsigned long long attribute_used __attribute__((aligned(8))) one_null = 0xffff0000ffff0000ULL;
-unsigned long __attribute__((aligned(16))) costab_mmx[] =
+unsigned int __attribute__((aligned(16))) costab_mmx[] =
 {
 	1056974725,
 	1057056395,
